# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pershing307 Member Portal API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface BootstrapStatus {
  bootstrapped: boolean;
}

export interface BootstrapInput {
  /**
     * @minLength 2
     * @maxLength 200
     */
  lodgeName: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  lodgeNumber: string;
  timezone: string;
  adminEmail: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  adminFirstName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  adminLastName: string;
  /** @minLength 12 */
  adminPassword: string;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: string | null;
  /** @nullable */
  smtpUser?: string | null;
  /** @nullable */
  smtpFromEmail?: string | null;
  /** @nullable */
  smtpFromName?: string | null;
}

export interface LodgeSummary {
  id: string;
  name: string;
  number: string;
}

export interface BootstrapResult {
  success: boolean;
  message: string;
  lodge: LodgeSummary;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface TwoFactorInput {
  /**
     * @minLength 6
     * @maxLength 8
     */
  code: string;
}

export interface TwoFactorStatus {
  enabled: boolean;
  /** @nullable */
  enrolledAt?: string | null;
  hasPendingEnrollment?: boolean;
}

export interface TwoFactorEnrollResult {
  qrCodeUri: string;
  backupCodes: string[];
}

export interface TwoFactorVerifyEnrollInput {
  /**
     * @minLength 6
     * @maxLength 8
     */
  code: string;
}

export interface ChangePasswordInput {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 12 */
  newPassword: string;
}

export interface ProfileUpdateInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  firstName?: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  lastName?: string;
  /**
     * @maxLength 100
     * @nullable
     */
  displayName?: string | null;
}

export interface UserRole {
  name: string;
  slug: string;
  permissionLevel: number;
}

export interface AuthUser {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  displayName?: string | null;
  mustChangePassword?: boolean;
  roles: UserRole[];
}

export interface LoginResult {
  requiresTwoFactor?: boolean;
  user?: AuthUser;
}

export type CurrentUserResultUser = AuthUser & ({
  /** @nullable */
  membershipStatus?: string | null;
});

export interface CurrentUserResult {
  user: CurrentUserResultUser;
}

export interface SuccessResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  /** @minLength 1 */
  token: string;
  /** @minLength 12 */
  password: string;
}

export interface UserSummary {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  membershipStatus?: string | null;
  isActive: boolean;
  /** @nullable */
  lastLoginAt?: string | null;
  createdAt: string;
}

export interface UserListResult {
  users: UserSummary[];
  total: number;
  limit: number;
  offset: number;
}

export interface RoleDetail {
  id: string;
  name: string;
  slug: string;
  permissionLevel: number;
}

export interface UserDetail {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  membershipStatus?: string | null;
  isActive: boolean;
  emailVerified: boolean;
  mustChangePassword: boolean;
  /** @nullable */
  lastLoginAt?: string | null;
  createdAt: string;
  roles: RoleDetail[];
}

export interface UserDetailResult {
  user: UserDetail;
}

export interface RoleAssignment {
  /** @minLength 1 */
  roleId: string;
}

export interface ProtectedDomain {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface DomainListResult {
  domains: ProtectedDomain[];
}

export interface UserDomainGrant {
  domainId: string;
  domainName: string;
  domainSlug: string;
  grantedAt: string;
}

export interface UserDomainAccessListResult {
  domains: UserDomainGrant[];
}

export interface DomainGrantInput {
  /** @minLength 1 */
  domainId: string;
}

export interface UserDegree {
  id: string;
  degree: number;
  degreeName: string;
  /** @nullable */
  conferredOn?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface UserDegreeListResult {
  degrees: UserDegree[];
}

export interface DegreeInput {
  /** @minimum 1 */
  degree: number;
  /** @nullable */
  conferredOn?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface DegreeDefinition {
  degree: number;
  name: string;
  abbreviation: string;
}

export interface DegreeDefinitionListResult {
  definitions: DegreeDefinition[];
}

export interface DegreeDefinitionUpdateInput {
  definitions: DegreeDefinition[];
}

export interface ConfigEntry {
  key: string;
  /** @nullable */
  value: string | null;
  description: string;
  isReadOnly: boolean;
}

export interface ConfigListResult {
  config: ConfigEntry[];
}

export interface ConfigUpdateInput {
  value: string;
}

export interface InvitationSummary {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  expiresAt: string;
  /** @nullable */
  acceptedAt?: string | null;
  /** @nullable */
  revokedAt?: string | null;
  createdAt: string;
}

export interface InvitationListResult {
  invitations: InvitationSummary[];
  smtpConfigured: boolean;
}

export interface InvitationInput {
  email: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  lastName: string;
  /** @nullable */
  roleId?: string | null;
}

export interface InvitationCreated {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  expiresAt: string;
}

export interface InvitationCreatedResult {
  invitation: InvitationCreated;
  smtpConfigured: boolean;
}

export interface InvitationLookup {
  email: string;
  firstName: string;
  lastName: string;
  lodgeName: string;
}

export interface InvitationLookupResult {
  invitation: InvitationLookup;
}

export interface AcceptInvitationInput {
  /** @minLength 1 */
  token: string;
  /** @minLength 12 */
  password: string;
}

export interface InvitationLinkResult {
  link: string;
}

export interface Role {
  id: string;
  name: string;
  slug: string;
  permissionLevel: number;
  isSystem: boolean;
}

export interface RoleListResult {
  roles: Role[];
}

/**
 * @nullable
 */
export type AuditLogEntryDetail = { [key: string]: unknown } | null;

export interface AuditLogEntry {
  id: string;
  /** @nullable */
  actorId?: string | null;
  /** @nullable */
  actorEmail?: string | null;
  action: string;
  /** @nullable */
  targetType?: string | null;
  /** @nullable */
  targetId?: string | null;
  /** @nullable */
  detail?: AuditLogEntryDetail;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  createdAt: string;
}

export interface AuditLogResult {
  logs: AuditLogEntry[];
  limit: number;
  offset: number;
}

export type ListUsersParams = {
/**
 * @minimum 1
 * @maximum 200
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
search?: string;
};

export type ListAuditLogsParams = {
limit?: number;
offset?: number;
};