# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Pershing307 Member Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful or 2FA required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
        "423":
          description: Account locked
  /auth/login/2fa:
    post:
      operationId: verifyTwoFactor
      tags: [auth]
      summary: Complete login with 2FA code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorInput"
      responses:
        "200":
          description: 2FA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid code
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResult"
        "401":
          description: Not authenticated
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent (always succeeds to prevent enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
        "400":
          description: Invalid or expired token
...
  /auth/2fa:
    delete:
      operationId: disableTwoFactor
      tags: [auth]
      summary: Disable 2FA for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorInput"
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
        "401":
          description: Invalid code
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password for authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
        "400":
          description: Invalid current password
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    TwoFactorInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 8
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 12
    ProfileUpdateInput:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        displayName:
          type: ["string", "null"]
          maxLength: 100
    UserRole:
      type: object
      required: [name, slug, permissionLevel]
      properties:
        name:
          type: string
        slug:
          type: string
        permissionLevel:
          type: integer
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, roles]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: ["string", "null"]
        mustChangePassword:
          type: boolean
        roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRole"
    LoginResult:
      type: object
      properties:
        requiresTwoFactor:
          type: boolean
        user:
          $ref: "#/components/schemas/AuthUser"
    CurrentUserResult:
      type: object
      required: [user]
      properties:
        user:
          allOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: object
              properties:
                membershipStatus:
                  type: ["string", "null"]
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 12
...