# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pershing307 Member Portal API
 * OpenAPI spec version: 0.2.0
 */

export interface UserRole {
  name: string;
  slug: string;
  permissionLevel: number;
}